Resolves #577 by shared_from_this callback rework
This issue manifested with random coredumps in the onRetryTimer()
callback in transport_publisher_link.cpp. These crashes were infrequent,
and extremely difficult to reproduce. They appeared to happen when
many Subscriptions were being destroyed, which made it likely to be
a destruction scheduling anomaly.

It turned out the issue was caused by a few functions that were attached
to boost::signals2::signals. These functions did not have shared_ptr's to
the classes that they were bound to, causing a coredump if the original
classes were destructed. This was particularly problematic if the class
was destroyed while the signal function was in the middle of execution.
Due to a mutex locking strategy in CallbackQueue::CallOneCB() causing semi-
random thread execution timing, this issue was only seen on occasion.

The solution to this bug was to rework how DropSignals were implemented.
Additional function parameters for tracked_objects were added to for
dropping connections in order to allow for an object's shared_from_this
shared pointer to be passed into the function that was boost::bind'ed to
a signal2::signal. This prevented the destruction of a Subscriber from
destroying the objects which thse Drop functions were acting upon.

Reproduction of this issue was incredibly difficult, since it required
the timing of the class destruction to perfectly coincide with the
execution of the callback function using that class. We found this could
only be reliably reproduced when running overnight simulations of hundreds
of Subscribers by Rethink Robotics Automated QA team. The occurance of
this bug was producing roughly 1 to 5 coredumps per night in this testing
environment. Since this patch was applied, zero coredumps relating to this
issue have been discovered over the course of a week of testing.

Major credit to Jeff Adams (@jeffadamsc) for pursuing the shared_from_this
callback connection strategy, and for QA allowing us to run these tests on
their nightly runs.